Added test cases for all piece moves.

+++ ChessEngineInCSharp/ChessEngine/Piece.cs
@@ -8,6 +8,6 @@ namespace ChessEngine
     {
         public string Name { get; set; }
         public bool IsWhite { get; set; }
-        public short Value { get; set; }
+        public int Value { get; set; }
     }
 }
